mv drs run localization into 'prepare'
